+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG
 
+## 0.0.2
+- Added: session token validation on middleware.
+- Added: Redirect on sign in.
+- Fixed: Infinite-loop on fetchArtefacts useEffect()
+- Fixed: Artefacts Table started on page 1 instead of page 0. 
+
 ## 0.0.1
 - First functional version of BAR UI.